+++ src/main/java/utils/Utils.java
@@ -63,8 +63,8 @@ public final class Utils {
 				new String[] { Configed.getResourceValue("buttonClose") }, 500, 300);
 
 		StringBuilder message = new StringBuilder();
-		message.append(Globals.APPNAME + "  " + Configed.getResourceValue("DPassword.version") + "  " + Globals.VERSION
-				+ " (" + Globals.VERDATE + ")");
+		message.append(Globals.APPNAME + "  " + Configed.getResourceValue("LoginDialog.version") + "  "
+				+ Globals.VERSION + " (" + Globals.VERDATE + ") ");
 		message.append("\n");
 		message.append("The opsi-logviewer is part of the " + Globals.APPNAME + " since version 4.2.22.1\n");
 		message.append("______________________________________________________________________\n");